Moving tests into tests/ root directory.
