+++ jest.config.js
@@ -1,4 +1,4 @@
-module.exports = /** @type {import('ts-jest/dist/types').InitialOptionsTsJest} */({
-  testMatch: ['**/out/**/*.test.js'],
+module.exports = {
+  testMatch: ['**/out/tests/**/*.test.js'],
   watchPathIgnorePatterns: ['src/'],
-});
+};
