+++ tests/deal-generator/test-cases/basics.test.ts
