+++ tests/deal-generator/test-cases/partner-transacted.test.ts
