+++ tests/deal-generator/test-cases/smoke.test.ts
@@ -1,4 +1,4 @@
-import { License } from "../../../../marketplace/model/license";
+import { License } from "../../../src/lib/marketplace/model/license";
 import { runDealGenerator } from "../utils";
 
 it(`Sets deal properties correctly`, () => {
