+++ tsconfig.json
@@ -21,6 +21,7 @@
   "include": [
     "src/**/*.js",
     "src/**/*.ts",
-    "src/**/*.d.ts"
+    "src/**/*.d.ts",
+    "tests/**/*.ts",
   ]
 }
